refactor(vm): MapPaneViewModel / PreviewPaneViewModel の IO 依存をサービス経由へ移管 (#92)
- MapPaneViewModel: GetPinPath (Path.Combine) → IMapPaneService.GetPinImagePath へ移管
- MapPaneViewModel: TryCreatePinStyle の File.Exists → IMapPaneService.FileExistsAtPath へ移管
- GetPinPath private static メソッドを削除、using System.IO を除去
- PreviewPaneViewModel: ExifService.GetMetadataAsync 直呼び出し → _service.GetMetadataAsync へ統一
- IPreviewPaneService / PreviewPaneService に GetMetadataAsync を追加
